See merge request Network_OM/PNTL!58

+++ PNTL_agent/src/AgentJsonAPI.cpp
@@ -13,17 +13,17 @@ using namespace boost::property_tree;
 {
 "ServerAntServer" :
     {
-        "IP"    : "127.0.0.1", 
+        "IP"    : "127.0.0.1",
         "Port"  : "2400"
     },
 "ServerAntCollector" :
     {
-        "IP"    : "127.0.0.1", 
+        "IP"    : "127.0.0.1",
         "Port"  : "2400"
     },
 "ServerAntAgent" :
     {
-        "IP"    : "10.78.221.45", 
+        "IP"    : "10.78.221.45",
         "Port"  : "2400"
     }
 }
@@ -32,21 +32,21 @@ using namespace boost::property_tree;
 INT32 ParserLocalCfg(const char * pcJsonData, ServerAntAgentCfg_C * pcCfg)
 {
     INT32 iRet = AGENT_OK;
-    
+
     // boost库中出现错误会抛出异常, 未被catch的异常会逐级上报, 最终导致进程abort退出.
     try
     {
         UINT32 uiIp, uiPort, uiData;
         string strTemp;
-        
+
         // pcData字符串转存stringstream格式, 方便后续boost::property_tree处理.
         stringstream ssStringData(pcJsonData);
 
         // boost::property_tree对象, 用于存储json格式数据.
         ptree ptDataRoot, ptDataTmp;
-        
+
         read_json(ssStringData, ptDataRoot);
-        
+
         // 解析LogCfg数据.
         ptDataTmp.clear();
         ptDataTmp = ptDataRoot.get_child("LogCfg");
@@ -57,7 +57,7 @@ INT32 ParserLocalCfg(const char * pcJsonData, ServerAntAgentCfg_C * pcCfg)
             JSON_PARSER_ERROR("SetNewLogDir failed[%d]", iRet);
             return iRet;
         }
-        
+
         // 刷新日志目录后, 再打印当前配置
         JSON_PARSER_INFO("JsonData[%s]", pcJsonData);
 
@@ -77,15 +77,15 @@ INT32 ParserLocalCfg(const char * pcJsonData, ServerAntAgentCfg_C * pcCfg)
         // 解析ServerAntCollector数据.
         ptDataTmp.clear();
         ptDataTmp = ptDataRoot.get_child("ServerAntCollector");
-        
+
         strTemp = ptDataTmp.get<string>("Protocol");
-        
+
 
         // 解析ServerAntAgent数据.
         ptDataTmp.clear();
         ptDataTmp = ptDataRoot.get_child("ServerAntAgent");
 		strTemp = ptDataTmp.get<string>("MgntIP");
-		
+
 		uiIp = sal_inet_aton(strTemp.c_str());
         iRet = pcCfg->SetMgntIP(uiIp);
 		if (iRet)
@@ -101,7 +101,7 @@ INT32 ParserLocalCfg(const char * pcJsonData, ServerAntAgentCfg_C * pcCfg)
             JSON_PARSER_ERROR("SetHostname failed[%d]", iRet);
             return iRet;
         }
-		
+
         strTemp = ptDataTmp.get<string>("AgentIP");
         uiIp = sal_inet_aton(strTemp.c_str());
         uiPort = ptDataTmp.get<UINT32>("Port");
@@ -172,7 +172,7 @@ INT32 ParserLocalCfg(const char * pcJsonData, ServerAntAgentCfg_C * pcCfg)
         JSON_PARSER_ERROR("Caught exception [%s] when ParserLocalCfg. LocalCfg:[%s]", e.what(), pcJsonData);
         return AGENT_E_ERROR;
     }
-    
+
     return iRet;
 }
 
@@ -201,7 +201,7 @@ INT32 CreatProbeListRequestPostData(ServerAntAgentCfg_C * pcCfg, stringstream * 
         ptDataRoot.put("Action", NormalFlowRequestAction);
         ptDataRoot.put("target", "ServerTopology");
         ptDataRoot.put("scope", "global");
-        
+
         UINT32 uiIp;
         iRet = pcCfg->GetAgentAddress(&uiIp, NULL);
         if (iRet)
@@ -211,7 +211,7 @@ INT32 CreatProbeListRequestPostData(ServerAntAgentCfg_C * pcCfg, stringstream * 
         }
         ptDataTemp.put("agent-ip", sal_inet_ntoa(uiIp));
         ptDataRoot.put_child("content", ptDataTemp);
-        
+
         write_json(ssJsonData, ptDataRoot);
         (*pssPostData) << ssJsonData.str();
     }
@@ -233,7 +233,7 @@ INT32 CreatAgentIPRequestPostData(ServerAntAgentCfg_C * pcCfg, stringstream * ps
     {
         stringstream ssJsonData;
         ptree ptDataRoot;
-        
+
         UINT32 uiIp, uiMgntIp;
         iRet = pcCfg->GetAgentAddress(&uiIp, NULL);
         if (iRet)
@@ -243,8 +243,8 @@ INT32 CreatAgentIPRequestPostData(ServerAntAgentCfg_C * pcCfg, stringstream * ps
         }
 
 		iRet = pcCfg->GetMgntIP(&uiMgntIp);
-		
-        ptDataRoot.put("vbond-ip", sal_inet_ntoa(uiIp));    // 数据面IP 
+
+        ptDataRoot.put("vbond-ip", sal_inet_ntoa(uiIp));    // 数据面IP
 		ptDataRoot.put("agent-ip", sal_inet_ntoa(uiMgntIp));
 		ptDataRoot.put("hostname", pcCfg->GetHostname().c_str());
         write_json(ssJsonData, ptDataRoot);
@@ -262,9 +262,9 @@ INT32 CreatAgentIPRequestPostData(ServerAntAgentCfg_C * pcCfg, stringstream * ps
 /*
 {
     "orgnizationSignature": "HuaweiDC3ServerAntsFull",
-    "flow": 
+    "flow":
     {
-        
+
         "sip": "",
         "dip": "",
         "sport": "",
@@ -272,23 +272,23 @@ INT32 CreatAgentIPRequestPostData(ServerAntAgentCfg_C * pcCfg, stringstream * ps
         "ip-protocol": "icmp:udp",
         "dscp": "",
         "urgent-flag": "0:1",
-        "topology-tag": 
+        "topology-tag":
         {
             "level": "",
             "svid": "",
             "dvid": ""
         },
-        "time": 
+        "time":
         {
             "t1": "",
             "t2": "",
             "t3": "",
             "t4": ""
         },
-        "statistics": 
+        "statistics":
         {
             "packet-sent": "",
-            "packet-drops": "",            
+            "packet-drops": "",
             "50percentile": "",
             "99percentile": ""
             "standard-deviation": ""
@@ -308,15 +308,15 @@ INT32 CreatLatencyReportData(AgentFlowTableEntry_S * pstAgentFlowEntry, stringst
     {
         ptree ptDataRoot, ptDataFlowArray, ptDataFlowEntry;
         ptree ptDataFlowEntryTemp;
-        
+
         stringstream ssJsonData;
 
         ptDataRoot.clear();
         ptDataRoot.put("orgnizationSignature", LatencyReportSignature);
-        
+
         // 清空Flow Entry Array
         ptDataFlowArray.clear();
-        
+
         // 生成一个Flow Entry的数据
         {
             ptDataFlowEntry.clear();
@@ -356,9 +356,9 @@ INT32 CreatLatencyReportData(AgentFlowTableEntry_S * pstAgentFlowEntry, stringst
             ptDataFlowEntryTemp.put("sid", pstAgentFlowEntry->stFlowKey.stServerTopo.uiSvid);
             ptDataFlowEntryTemp.put("did", pstAgentFlowEntry->stFlowKey.stServerTopo.uiDvid);
 #endif
-            
+
             ptDataFlowEntry.put_child("topology-tag", ptDataFlowEntryTemp);
-            
+
             // 处理time信息
             ptDataFlowEntryTemp.clear();
             ptDataFlowEntryTemp.put("t1", pstAgentFlowEntry->stFlowDetectResult.lT1);
@@ -368,11 +368,11 @@ INT32 CreatLatencyReportData(AgentFlowTableEntry_S * pstAgentFlowEntry, stringst
             ptDataFlowEntry.put_child("time", ptDataFlowEntryTemp);
 
             // 处理statistics信息
-            ptDataFlowEntryTemp.clear();        
+            ptDataFlowEntryTemp.clear();
             ptDataFlowEntryTemp.put("packet-sent", pstAgentFlowEntry->stFlowDetectResult.lPktSentCounter);
             ptDataFlowEntryTemp.put("packet-drops", pstAgentFlowEntry->stFlowDetectResult.lPktDropCounter);
             ptDataFlowEntryTemp.put("50percentile", pstAgentFlowEntry->stFlowDetectResult.lLatency50Percentile);
-            ptDataFlowEntryTemp.put("99percentile", pstAgentFlowEntry->stFlowDetectResult.lLatency99Percentile);            
+            ptDataFlowEntryTemp.put("99percentile", pstAgentFlowEntry->stFlowDetectResult.lLatency99Percentile);
             ptDataFlowEntryTemp.put("standard-deviation", pstAgentFlowEntry->stFlowDetectResult.lLatencyStandardDeviation);
             ptDataFlowEntryTemp.put("min", pstAgentFlowEntry->stFlowDetectResult.lLatencyMin);
             ptDataFlowEntryTemp.put("max", pstAgentFlowEntry->stFlowDetectResult.lLatencyMax);
@@ -388,7 +388,7 @@ INT32 CreatLatencyReportData(AgentFlowTableEntry_S * pstAgentFlowEntry, stringst
 
         ssJsonData.clear();
         ssJsonData.str("");
-        write_json(ssJsonData, ptDataRoot);    
+        write_json(ssJsonData, ptDataRoot);
         (*pssReportData) << ssJsonData.str();
     }
     catch (exception const & e)
@@ -437,7 +437,7 @@ INT32 CreatDropReportData(AgentFlowTableEntry_S * pstAgentFlowEntry, stringstrea
     {
         ptree ptDataRoot, ptDataFlowArray, ptDataFlowEntry;
         ptree ptDataFlowEntryTemp;
-        
+
         stringstream ssJsonData;
 
         ptDataRoot.clear();
@@ -449,7 +449,7 @@ INT32 CreatDropReportData(AgentFlowTableEntry_S * pstAgentFlowEntry, stringstrea
         // 生成一个Flow Entry的数据
         {
             ptDataFlowEntry.clear();
-            
+
             ptDataFlowEntry.put("sip", sal_inet_ntoa(pstAgentFlowEntry->stFlowKey.uiSrcIP));
             ptDataFlowEntry.put("dip", sal_inet_ntoa(pstAgentFlowEntry->stFlowKey.uiDestIP));
             ptDataFlowEntry.put("sport", pstAgentFlowEntry->stFlowKey.uiSrcPort);
@@ -472,7 +472,7 @@ INT32 CreatDropReportData(AgentFlowTableEntry_S * pstAgentFlowEntry, stringstrea
             ptDataFlowEntry.put("dscp",pstAgentFlowEntry->stFlowKey.uiDscp);
             ptDataFlowEntry.put("urgent-flag",pstAgentFlowEntry->stFlowKey.uiUrgentFlow);
 
-            // 处理topology-tag信息        
+            // 处理topology-tag信息
             ptDataFlowEntryTemp.clear();
             ptDataFlowEntryTemp.put("level", pstAgentFlowEntry->stFlowKey.stServerTopo.uiLevel);
 #if 0
@@ -492,9 +492,9 @@ INT32 CreatDropReportData(AgentFlowTableEntry_S * pstAgentFlowEntry, stringstrea
             ptDataFlowEntryTemp.clear();
             ptDataFlowEntryTemp.put("t", pstAgentFlowEntry->stFlowDetectResult.lT5);
             ptDataFlowEntryTemp.put("packet-sent", pstAgentFlowEntry->stFlowDetectResult.lPktSentCounter);
-            ptDataFlowEntryTemp.put("packet-drops", pstAgentFlowEntry->stFlowDetectResult.lPktDropCounter);       
+            ptDataFlowEntryTemp.put("packet-drops", pstAgentFlowEntry->stFlowDetectResult.lPktDropCounter);
             ptDataFlowEntry.put_child("statistics", ptDataFlowEntryTemp);
-            
+
             // 加入json数组
             ptDataFlowArray.push_back(make_pair("", ptDataFlowEntry));
         }
@@ -504,7 +504,7 @@ INT32 CreatDropReportData(AgentFlowTableEntry_S * pstAgentFlowEntry, stringstrea
 
         ssJsonData.clear();
         ssJsonData.str("");
-        write_json(ssJsonData, ptDataRoot);    
+        write_json(ssJsonData, ptDataRoot);
         (*pssReportData) << ssJsonData.str();
     }
     catch (exception const & e)
@@ -519,20 +519,20 @@ INT32 CreatDropReportData(AgentFlowTableEntry_S * pstAgentFlowEntry, stringstrea
 
 /*
 ServerAntServer 下发的紧急探测流格式
-post 
+post
 key = ServerAntAgent
-data = 
+data =
 {
     "orgnizationSignature": "HuaweiDC3ServerAntsProbelistIssue",
     "serverIP": "10.1.1.1",
     "action": "post",
     "content": "probe-list",
-    
+
     "flow": [
         {
-        "urgent": "true:false",        
+        "urgent": "true:false",
         "sip": "",
-        "dip": "", 
+        "dip": "",
         "ip-protocol": "icmp:tcp:udp",
         "sport-min": "",
         "sport-max": "",
@@ -545,9 +545,9 @@ data =
         },
 
         {
-        "urgent": "true:false",        
+        "urgent": "true:false",
         "sip": "",
-        "dip": "", 
+        "dip": "",
         "ip-protocol": "icmp:tcp:udp",
         "sport-min": "",
         "sport-max": "",
@@ -577,7 +577,7 @@ INT32 GetFlowInfoFromJsonFlowEntry(ptree ptFlowEntry, ServerFlowKey_S * pstNewSe
 
         // 初始化
         sal_memset(pstNewServerFlowKey, 0, sizeof(ServerFlowKey_S));
-        
+
         // 解析Urgent
 #if 0
         pstNewServerFlowKey->uiUrgentFlow   = ptFlowEntry.get<UINT32>("urgent-flag");
@@ -588,11 +588,11 @@ INT32 GetFlowInfoFromJsonFlowEntry(ptree ptFlowEntry, ServerFlowKey_S * pstNewSe
         else
             pstNewServerFlowKey->uiUrgentFlow = 0;
 #endif
-        
+
         // 解析Protocol
         strTemp = ptFlowEntry.get<string>("ip-protocol");
         if (0 == strTemp.compare("udp"))
-            pstNewServerFlowKey->eProtocol = AGENT_DETECT_PROTOCOL_UDP;                
+            pstNewServerFlowKey->eProtocol = AGENT_DETECT_PROTOCOL_UDP;
         else if (0 == strTemp.compare("tcp"))
             pstNewServerFlowKey->eProtocol = AGENT_DETECT_PROTOCOL_TCP;
         else if (0 == strTemp.compare("icmp"))
@@ -602,7 +602,7 @@ INT32 GetFlowInfoFromJsonFlowEntry(ptree ptFlowEntry, ServerFlowKey_S * pstNewSe
 
         strTemp = ptFlowEntry.get<string>("sip");
         pstNewServerFlowKey->uiSrcIP = sal_inet_aton(strTemp.c_str());
-        
+
         strTemp = ptFlowEntry.get<string>("dip");
         pstNewServerFlowKey->uiDestIP = sal_inet_aton(strTemp.c_str());
 
@@ -616,18 +616,18 @@ INT32 GetFlowInfoFromJsonFlowEntry(ptree ptFlowEntry, ServerFlowKey_S * pstNewSe
             //pstNewServerFlowKey->stServerTopo.uiSvid   = ptFlowEntryTopo.get<UINT32>("svid");
             uiDataTemp = 0;
             strTemp = ptFlowEntryTopo.get<string>("svid");
-            sscanf(strTemp.c_str(), "0x%x", &uiDataTemp); 
+            sscanf(strTemp.c_str(), "0x%x", &uiDataTemp);
             pstNewServerFlowKey->stServerTopo.uiSvid   = uiDataTemp;
             //pstNewServerFlowKey->stServerTopo.uiDvid   = ptFlowEntryTopo.get<UINT32>("dvid");
             uiDataTemp = 0;
             strTemp = ptFlowEntryTopo.get<string>("dvid");
-            sscanf(strTemp.c_str(), "0x%x", &uiDataTemp); 
+            sscanf(strTemp.c_str(), "0x%x", &uiDataTemp);
             pstNewServerFlowKey->stServerTopo.uiDvid   = uiDataTemp;
         #else
             pstNewServerFlowKey->stServerTopo.uiSvid   = ptFlowEntryTopo.get<UINT32>("src-id");
             pstNewServerFlowKey->stServerTopo.uiDvid   = ptFlowEntryTopo.get<UINT32>("dst-id");
         #endif
-        
+
         pstNewServerFlowKey->stServerTopo.uiLevel  = ptFlowEntryTopo.get<UINT32>("level");
     }
     catch (exception const & e)
@@ -651,7 +651,7 @@ INT32 GetFlowInfoFromJsonFlowEntry(ptree ptFlowEntry, ServerFlowKey_S * pstNewSe
 INT32 IssueFlowFromJsonFlowArray(ptree ptFlowArray, FlowManager_C* pcFlowManager, UINT32 uiIsUrgentFlow)
 {
     INT32 iRet = AGENT_OK;
-    
+
     // boost库中出现错误会抛出异常, 未被catch的异常会逐级上报, 最终导致进程abort退出.
     try
     {
@@ -723,7 +723,7 @@ INT32 ProcessUrgentFlowFromServer(const char * pcJsonData, FlowManager_C* pcFlow
         JSON_PARSER_ERROR("Caught exception [%s] when ProcessUrgentFlowFromServer. Flow info[%s]", e.what(), pcJsonData);
         return AGENT_E_ERROR;
     }
-    
+
     return iRet;
 }
 
@@ -737,12 +737,12 @@ ServerAntServer 回复的普通探测流格式
     "serverIP": "10.1.1.1",
     "action": "reply",
     "content": "probe-list",
-    
+
     "flow": [
         {
-        "urgent": "true:false",        
+        "urgent": "true:false",
         "sip": "",
-        "dip": "", 
+        "dip": "",
         "ip-protocol": "icmp:tcp:udp",
         "sport-min": "",
         "sport-max": "",
@@ -755,9 +755,9 @@ ServerAntServer 回复的普通探测流格式
         },
 
         {
-        "urgent": "true:false",        
+        "urgent": "true:false",
         "sip": "",
-        "dip": "", 
+        "dip": "",
         "ip-protocol": "icmp:tcp:udp",
         "sport-min": "",
         "sport-max": "",
@@ -784,7 +784,7 @@ INT32 ProcessNormalFlowFromServer(char * pcJsonData, FlowManager_C* pcFlowManage
     // boost库中出现错误会抛出异常, 未被catch的异常会逐级上报, 最终导致进程abort退出.
     try
     {
-    
+
         // pcData字符串转存stringstream格式, 方便后续boost::property_tree处理.
         stringstream ssStringData(pcJsonData);
 
@@ -796,12 +796,12 @@ INT32 ProcessNormalFlowFromServer(char * pcJsonData, FlowManager_C* pcFlowManage
 #if 0
         // 检查Signature
         string strSignature =   ptDataRoot.get<string>("MessageSignature");
-        
+
         // 校验签名
         if (0 == strSignature.compare(NormalFlowReplaySignature))
         {
             string strAction    =   ptDataRoot.get<string>("Action");
-            
+
             // 校验action和content
             if( (0 != strAction.compare(NormalFlowReplayAction)) )
             {
@@ -809,7 +809,7 @@ INT32 ProcessNormalFlowFromServer(char * pcJsonData, FlowManager_C* pcFlowManage
                 // 校验失败
                 return AGENT_E_ERROR;
             }
-            
+
             // 从data中解析数据,填充stServerFlowKey, 然后调用FlowManager接口添加探测流.
             ptFlowArray.clear();
             ptFlowArray = ptDataRoot.get_child("flow");
@@ -819,12 +819,12 @@ INT32 ProcessNormalFlowFromServer(char * pcJsonData, FlowManager_C* pcFlowManage
                 JSON_PARSER_ERROR("Issue Flow From Json Flow Array failed [%d]", iRet);
                 return iRet;
             }
-                    
+
             return iRet;
         }
         else
         {
-            JSON_PARSER_WARNING("Unsupported Signature:[%s]", strSignature.c_str());        
+            JSON_PARSER_WARNING("Unsupported Signature:[%s]", strSignature.c_str());
             return AGENT_E_PARA;
         }
 #else
@@ -837,7 +837,7 @@ INT32 ProcessNormalFlowFromServer(char * pcJsonData, FlowManager_C* pcFlowManage
             JSON_PARSER_ERROR("Issue Flow From Json Flow Array failed [%d]. Flow info[%s]", iRet, pcJsonData);
             return iRet;
         }
-#endif  
+#endif
     }
 
     catch (exception const & e)
@@ -865,7 +865,7 @@ INT32 ProcessActionFlowFromServer(const char * pcJsonData, FlowManager_C* pcFlow
     // boost::property_tree对象, 用于存储json格式数据.
     ptree ptDataRoot;
 	UINT32 interval;
-    try 
+    try
 	{
         // 防止Json消息体不规范
         read_json(ssStringData, ptDataRoot);
@@ -877,7 +877,7 @@ INT32 ProcessActionFlowFromServer(const char * pcJsonData, FlowManager_C* pcFlow
         JSON_PARSER_ERROR("Parse Json message[%s] error [%s].", pcJsonData, e.what());
         return AGENT_E_ERROR;
 	}
-	INT32 iRet = pcFlowManager -> FlowManagerAction(interval);
+	INT32 iRet = pcFlowManager -> FlowManagerAction((INT32)interval);
 	return iRet;
 }
 
@@ -892,8 +892,8 @@ INT32 ProcessActionFlowFromServer(const char * pcJsonData, FlowManager_C* pcFlow
 		"dscp":"",
 		"lossPkg_timeout":"",
 		"package_rate":""
-		
-		
+
+
 	}
 */
 INT32 ProcessConfigFlowFromServer(const char * pcJsonData, FlowManager_C* pcFlowManager)
@@ -905,7 +905,7 @@ INT32 ProcessConfigFlowFromServer(const char * pcJsonData, FlowManager_C* pcFlow
     ptree ptDataRoot;
 	UINT32 interval;
 	INT32 iRet = AGENT_OK;
-    try 
+    try
 	{
         // 防止Json消息体不规范
         read_json(ssStringData, ptDataRoot);
@@ -916,7 +916,7 @@ INT32 ProcessConfigFlowFromServer(const char * pcJsonData, FlowManager_C* pcFlow
 		    JSON_PARSER_ERROR("SetDectPeriod[%u] failed[%d], range should be in [%u, %u]", interval, iRet, MIN_PROBE_PERIOD, MAX_PROBE_PERIOD);
 		    return AGENT_E_ERROR;
 		}
-		
+
 		interval = ptDataRoot.get<UINT32>("port_count");
 		iRet = pcFlowManager->pcAgentCfg->SetPortCount(interval);
 		if (iRet)
@@ -940,7 +940,7 @@ INT32 ProcessConfigFlowFromServer(const char * pcJsonData, FlowManager_C* pcFlow
 		    JSON_PARSER_ERROR("SetMaxDelay[%u] failed[%d]");
 		    return AGENT_E_ERROR;
 		}
-		
+
 		interval = ptDataRoot.get<UINT32>("dscp");
 		iRet = pcFlowManager->pcAgentCfg->SetDscp(interval);
 		if (iRet)
